# src/hardware/Omniscan/omniscan-orchestrator/openapi.yaml
openapi: 3.0.3
info:
  title: Omniscan Orchestrator API
  description: |
    Medical device orchestration API for Omniscan XRD system.
    
    **Privacy-First Architecture**: Patient PII never leaves the orchestrator database - only UUIDs are transmitted to the hardware server and cloud.
    
    **Safety-Critical**: Physical Enable Button confirmation required for potentially harmful operations (X-ray activation, motion system).
    
    **Medical Device Compliance**: Complete audit trails with cryptographic hash chaining for regulatory compliance (FDA 21 CFR Part 11, HIPAA, GDPR).
    
    ## Authentication
    All endpoints (except `/api/auth/login`, `/api/connection/status`, `/api/debug/status`, and Command Discovery endpoints) require authentication via the `X-Session-Id` header.
    
    ## Safety Requirements
    - **Enable Button Operations**: Initialize detector, initialize motion, start exposure, motion commands require physical Enable Button confirmation (20-second window)
    - **Safe Operations**: Stop operations, power off, read states do not require Enable Button
    - **Safety Interlocks**: Key switch ON, emergency stop not pressed, door closed, cooling OK, power OK
  version: 0.1.0
  contact:
    name: Omniscan Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://localhost:8081
    description: Alternate local server
...
paths:
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Login user
      description: |
        Authenticate user with password and key switch verification.
        
        Process:
        1. Verify username/password (local DB or LDAP)
        2. Check hardware key switch is ON (via gRPC)
        3. Create session if both pass
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login response (check success field)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
  /api/auth/logout:
    post:
      tags:
        - Authentication
      summary: Logout user
      description: Logout user and invalidate session
      operationId: logout
      security:
        - SessionAuth: []
      responses:
        '200':
          description: Logout response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/state:
    get:
      tags:
        - System Health
      summary: Get system state
      description: |
        Get current system state for UI polling.
        
        Returns essential state info:
        - state: Current server state (IDLE, PENDING_ARMED, RUNNING, etc.)
        - interlocks: Safety interlock status
        - devices: Device status (detector, motion, PDU, GPIO)
        - calibration: Calibration validity
      operationId: getSystemState
      security:
        - SessionAuth: []
      responses:
        '200':
          description: System state
          content:
            application/json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                    enum: [IDLE, PENDING_ARMED, RUNNING, STOPPING, SAFE, LOCKED, CALIBRATION, MAINTENANCE]
                  timestamp:
                    type: string
                    format: date-time
                  interlocks:
                    $ref: '#/components/schemas/SafetyInterlocks'
                  devices:
                    type: object
                    properties:
                      pdu:
                        $ref: '#/components/schemas/PduStatus'
                      gpio:
                        $ref: '#/components/schemas/GpioStatus'
                      detector:
                        $ref: '#/components/schemas/DetectorStatus'
                      motion:
                        $ref: '#/components/schemas/MotionStatus'
                  calibration:
                    type: object
                    properties:
                      valid:
                        type: boolean
                      expires_at:
                        type: string
                        format: date-time
                        nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /api/gpio/enable-button:
    get:
      tags:
        - GPIO & Safety
      summary: Check Enable Button
      description: Check Enable Button status
      operationId: checkEnableButton
      security:
        - SessionAuth: []
      responses:
        '200':
          description: Enable Button status
          content:
            application/json:
              schema:
                type: object
                properties:
                  active:
                    type: boolean
                    description: Whether button is currently active
                  remaining_secs:
                    type: integer
                    nullable: true
                    description: Seconds remaining in 20s window
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-Id
      description: Session ID obtained from login
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: User identifier
        password:
          type: string
          format: password
          description: User password
    LoginResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
        session_id:
          type: string
          nullable: true
        user_id:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
          enum: [operator, engineer, administrator]
        error:
          type: string
          nullable: true
    LogoutResponse:
      type: object
      required:
        - success
        - message
      properties:
        success:
          type: boolean
        message:
          type: string
...
    SafetyInterlocks:
      type: object
      required:
        - overall_safe
        - key_switch
        - enable_button
        - door_closed
        - emergency_stop
        - radiation_safe
        - cooling_ok
        - power_ok
      properties:
        overall_safe:
          type: boolean
        key_switch:
          type: boolean
        enable_button:
          type: boolean
        door_closed:
          type: boolean
        emergency_stop:
          type: boolean
        radiation_safe:
          type: boolean
        cooling_ok:
          type: boolean
        power_ok:
          type: boolean
    DetectorStatus:
      type: object
      required:
        - powered
        - status
      properties:
        powered:
          type: boolean
        status:
          type: string
          enum: [OFF, IDLE, READY, INIT, EXPOSING, READING, ERROR]
        temperature:
          type: number
          format: float
          nullable: true
        voltage:
          type: number
          format: float
          nullable: true
    MotionStatus:
      type: object
      required:
        - powered
        - status
      properties:
        powered:
          type: boolean
        status:
          type: string
          enum: [OFF, IDLE, READY, INIT, MOVING, HOMING, ERROR, LIMIT_HIT]
        is_homed:
          type: boolean
          nullable: true
        position:
          type: number
          format: float
          nullable: true
    GpioStatus:
      type: object
      required:
        - powered
        - status
      properties:
        powered:
          type: boolean
        status:
          type: string
          enum: [Active, Off]
    PduStatus:
      type: object
      required:
        - powered
        - status
      properties:
        powered:
          type: boolean
        status:
          type: string
          enum: [Active, Off]
        outputs:
          type: object
          additionalProperties:
            type: boolean
          nullable: true
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        code:
          type: string
          nullable: true
        details:
          type: string
          nullable: true
  responses:
...
    Unauthorized:
      description: Unauthorized - missing or invalid session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServiceUnavailable:
      description: Service unavailable - hardware server not responding
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'